+++ application/controllers/payment.php
@@ -32,6 +32,10 @@ class Payment extends CI_Controller {
 	 */
 	public function index($payment_id=0)
 	{
+		$permission 	=	$this->module_model->get_permission_by_module_id_and_user_id(6, $this->session->userdata('user_id')); // module_id for bank is 20.....
+		if($permission->permission_add != 1){
+			redirect('access_control/denied/payment/add_payment','refresh');
+		}
 		$data							=	array();
 		$data['page_title']				=	"Inventory Management";
 
@@ -65,6 +69,10 @@ class Payment extends CI_Controller {
 
 	public function view_payments()
 	{
+		$permission 	=	$this->module_model->get_permission_by_module_id_and_user_id(6, $this->session->userdata('user_id')); // module_id for bank is 20.....
+		if($permission->permission_view != 1){
+			redirect('access_control/denied/payment/all_payments','refresh');
+		}
 		$data							=	array();
 		$data['page_title']				=	"Inventory Management";
 		$nav_data['dev_key']			=	"payment";
@@ -85,6 +93,10 @@ class Payment extends CI_Controller {
 
 	public function add_payment()
 	{
+		$permission 	=	$this->module_model->get_permission_by_module_id_and_user_id(6, $this->session->userdata('user_id')); // module_id for bank is 20.....
+		if($permission->permission_add != 1){
+			redirect('access_control/denied/payment/add_payment','refresh');
+		}
 		$payment_data						=	array();
 		$payment_data['user_id']			=	$this->session->userdata('user_id');
 		$payment_data['user_name']			=	$this->session->userdata('user_name');
@@ -109,6 +121,10 @@ class Payment extends CI_Controller {
 
 	public function update_payment($payment_id)
 	{
+		$permission 	=	$this->module_model->get_permission_by_module_id_and_user_id(6, $this->session->userdata('user_id')); // module_id for bank is 20.....
+		if($permission->permission_edit != 1){
+			redirect('access_control/denied/payment/add_payment','refresh');
+		}
 		$payment_data						=	array();
 		$payment_data['user_id']			=	$this->session->userdata('user_id');
 		$payment_data['user_name']			=	$this->session->userdata('user_name');
@@ -133,94 +149,13 @@ class Payment extends CI_Controller {
 	}
 	public function delete_payment($payment_id)
 	{
+		$permission 	=	$this->module_model->get_permission_by_module_id_and_user_id(6, $this->session->userdata('user_id')); // module_id for bank is 20.....
+		if($permission->permission_delete != 1){
+			redirect('access_control/denied/payment/all_payments','refresh');
+		}
 		
 		$this->payment_model->delete_payment($payment_id);
 
 		redirect('payment/view_payments','refresh');
 	}
-
-
-
-
-	//---------------------VENDOR SECTION STARTS HERE
-	public function vendor($vendor_id=0)
-	{
-		$data						=	array();
-		$data['page_title']			=	"Inventory Management";
-		$nav_data['dev_key']		=	"vendor";
-		$nav_data['selected']		=	"add_vendor";
-		$nav_data['company_name']   =   $this->company_model->get_company_by_id(1)->company_name;
-		$nav_data['user_permission']=	$this->module_model->get_permission_by_user_id($this->session->userdata('user_id'));
-
-		$vendor_data				=	array();
-		if($vendor_id!=0){
-			$vendor_data['vendor']	=	$this->payment_model->get_vendor_by_id($vendor_id);	
-		}else{
-			$vendor_data['vendor']	=	NULL;
-		}
-		
-
-		$data['navigation']			=	$this->load->view('templates/navigation',$nav_data,TRUE);
-		$data['footer']				=	$this->load->view('templates/footer','',TRUE);
-		$data['content']			=	$this->load->view('partials/vendor',$vendor_data,TRUE);
-
-		$this->load->view('templates/main_template',$data);
-	}
-	public function view_vendors()
-	{
-		$data						=	array();
-		$data['page_title']			=	"Inventory Management";
-		$nav_data['dev_key']		=	"vendor";
-		$nav_data['selected']		=	"all_vendors";
-		$nav_data['company_name']   =   $this->company_model->get_company_by_id(1)->company_name;
-		$nav_data['user_permission']=	$this->module_model->get_permission_by_user_id($this->session->userdata('user_id'));
-
-		$vendor_data				=	array();
-		$vendor_data['vendor_list']	=	$this->payment_model->get_all_vendors();
-		$vendor_data['permission']	= 	$this->module_model->get_permission_by_module_id_and_user_id(4,$this->session->userdata('user_id'));
-
-		$data['navigation']			=	$this->load->view('templates/navigation',$nav_data,TRUE);
-		$data['footer']				=	$this->load->view('templates/footer','',TRUE);
-		$data['content']			=	$this->load->view('partials/vendor_list',$vendor_data,TRUE);
-
-		$this->load->view('templates/main_template',$data);
-	}
-
-	public function add_vendor()
-	{
-		$vendor_data					=	array();
-		$vendor_data['user_id']			=	$this->session->userdata('user_id');
-		$vendor_data['user_name']		=	$this->session->userdata('user_name');
-		$vendor_data['vendor_name']		=	$this->input->post('vendor_name','',TRUE);
-		$vendor_data['address']			=	$this->input->post('address','',TRUE);
-		$vendor_data['phone_no']		=	$this->input->post('phone_no','',TRUE);
-
-		$result							=	$this->payment_model->add_vendor($vendor_data);
-
-		redirect('payment/view_vendors','refresh');
-	}
-
-	public function update_vendor($vendor_id)
-	{
-		$vendor_data					=	array();
-		$vendor_data['user_id']			=	$this->session->userdata('user_id');
-		$vendor_data['user_name']		=	$this->session->userdata('user_name');
-		$vendor_data['vendor_name']		=	$this->input->post('vendor_name','',TRUE);
-		$vendor_data['address']			=	$this->input->post('address','',TRUE);
-		$vendor_data['phone_no']		=	$this->input->post('phone_no','',TRUE);
-
-		$result							=	$this->payment_model->update_vendor($vendor_data,$vendor_id);
-
-		redirect('payment/view_vendors','refresh');
-	}
-	public function delete_vendor($vendor_id)
-	{
-		
-		$this->payment_model->delete_vendor($vendor_id);
-
-		redirect('payment/view_vendors','refresh');
-	}
-
-	//---------------------VENDOR SECTION ENDS HERE
-
 }
